Post permissions

+++ app/Http/Controllers/PostController.php
@@ -5,7 +5,8 @@ namespace App\Http\Controllers;
 use App\Http\Requests\PostRequest;
 use App\Models\Post;
 use Exception;
-use Illuminate\Http\Request;
+
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Session;
 
 class PostController extends Controller
@@ -26,25 +27,37 @@ class PostController extends Controller
     }
     public function destroy(Post $post)
     {
-        try{
-            $post->comments()->delete();
-            $post->delete();
-            Session::flash('status', 'Post has been deleted');
-            return response()->json([
-                'status' => 'success'
-            ]);
-        } catch(Exception $e){
-            return response()->json([
-                'status' => 'error',
-                'message' => 'Something went wrong...'
-            ])->setStatusCode(500);
+        $user = Auth::user();
+
+        if($user->isAdmin() || $user->isModerator() || $user->isOwnerOfPost($post)){
+            try{
+                $post->comments()->delete();
+                $post->delete();
+                Session::flash('status', 'Post has been deleted');
+                return response()->json([
+                    'status' => 'success'
+                ]);
+            } catch(Exception $e){
+                return response()->json([
+                    'status' => 'error',
+                    'message' => 'Something went wrong...'
+                ])->setStatusCode(500);
+            }
+        } else {
+            abort(401, 'Unauthorized action.');
         }
     }
     public function edit(Post $post)
     {
-        return view('home.edit', compact(
-            'post'
-        ));
+        $user = Auth::user();
+        
+        if($user->isAdmin() || $user->isModerator() || $user->isOwnerOfPost($post)){
+            return view('home.edit', compact(
+                'post'
+            ));
+        } else {
+            abort(401, 'Unauthorized action.');
+        }
     }
     public function update(PostRequest $request, Post $post)
     {
